added optional ?connection=n parameter to add-reporter page
this is useful for converting PersisantConnection objects into Reporter
objects in the WebUI without any magic in the view. for when people
forget to register themselves.

note: this doesn't reassign any models which have been linked to the
PersistantConnection object to the new Reporter object. maybe that would
be cool.